tts/dangvansam-VietTTS-backend/requirements.txt
# TTS Backend Requirements - Using VietTTS compatible versions
# Yêu cầu TTS Backend - Sử dụng phiên bản tương thích VietTTS

# Base dependencies from VietTTS (working setup)
# Phụ thuộc cơ bản từ VietTTS (setup hoạt động)
conformer==0.3.2
diffusers==0.27.2
hydra-core==1.3.2
hyperpyyaml==1.2.2
librosa==0.10.2
omegaconf==2.3.0
onnx==1.16.0
onnxruntime-gpu==1.16.0  # GPU version for NVIDIA GPU support (matches viet-tts pyproject.toml)
protobuf==4.25
soundfile==0.12.1
torch==2.0.1
torchaudio==2.0.2
wget==3.2
vinorm>=2.0.7
underthesea>=8.3.0  # Vietnamese NLP (tokenization, sentence splitting) for Windows normalization
huggingface-hub==0.24.7  # Match viet-tts exactly (will patch diffusers to use hf_hub_download instead of cached_download)
silero-vad>=5.1.2
tiktoken>=0.8.0
openai-whisper>=20240930

# API & Web Framework / API và Framework Web
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
loguru>=0.7.2
click>=8.1.7

# Utilities / Tiện ích
numpy<2.0.0,>=1.21.6  # NumPy 2.x incompatible with onnxruntime 1.16.0 (compiled with NumPy 1.x)
safetensors>=0.4.0
tqdm>=4.65.0

# Optional: For Gradio interface / Tùy chọn: Cho giao diện Gradio
# gradio>=4.32.2

